+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Docs) Update Refs
 - (Feature) Expose Agency Shard Details
 - (Maintenance) Bump Go to 1.23.6, Kubernetes to 1.31.5 and other dependencies
+- (Feature) (Platform) Inventory Cluster details
 
 ## [1.2.44](https://github.com/arangodb/kube-arangodb/tree/1.2.44) (2025-02-03)
 - (Maintenance) Kubernetes 1.31.1 libraries